fix all issues with user repository and controller

+++ src/data/repositories/User.repository.js
@@ -26,22 +26,36 @@ export default class UserRepository {
             throw new Error('id is undefined');
         }
 
+        const userCheck = await this.findOne(user.id);
+
+        if(userCheck == undefined) {
+            return undefined;
+        }
+
         const userBO = user.toPersistenceObject();
+
         const result = await User.update(userBO, {
             where: {
                 id: user.id
             }
         });
+
         return this.findOne(user.id);
     };
 
     async delete(id) {
-        
+        const userCheck = await this.findOne(id);
+
+        if(userCheck == undefined) {
+            return undefined;
+        }
+
         const result = await User.destroy({
             where: {
                 id: id
             }
         });
+
         return result;
     }
 
@@ -52,6 +66,11 @@ export default class UserRepository {
                 id: id
             }
         });
+
+        if(result == null) {
+            return undefined;
+        };
+
         return new UserBO(
             result.dataValues.id, 
             result.dataValues.username, 
@@ -64,13 +83,17 @@ export default class UserRepository {
             result.dataValues.email, 
             result.dataValues.img_profile, 
             result.dataValues.birthdate
-            );
+        );
     };
                 
 
     async findAll() {
-        
         const result = await User.findAll();
+
+        if(result == null || result.length == 0) {
+            return undefined;
+        };
+
         return result;
     }
 }
